+++ front-end/src/app/components/admin/admin-patient/admin-patient.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { AdminService } from './../../../services//admin/admin.service';
+import swal from 'sweetalert';
 
 
 @Component({
@@ -35,19 +36,29 @@ export class AdminPatientComponent implements OnInit {
   }
 
   deletePatient(id) {
-    this.adminService.deletePatient(id).subscribe(res => {
-      if (res) {
-        alert('Patient Deleted successfully');
-        let index = this.patients.findIndex(obj => obj['_id'] == res['user']['_id']);
-        if (index !== -1) {
-          this.patients.splice(index, 1);
-        }
-      }
-    }, err => {
-      if (err) {
-        console.log(err);
-      }
+    swal({
+      title: "Are you sure?",
+      text: "Are you sure that you want to delete this patient?",
+      icon: "warning",
+      dangerMode: true,
     })
+      .then(willDelete => {
+        if (willDelete) {
+          this.adminService.deletePatient(id).subscribe(res => {
+            if (res) {
+              swal("", "Patient has been deleted!", "success");
+              let index = this.patients.findIndex(obj => obj['_id'] == res['user']['_id']);
+              if (index !== -1) {
+                this.patients.splice(index, 1);
+              }
+            }
+          }, err => {
+            if (err) {
+              console.log(err);
+            }
+          })
+        }
+      });
   }
 
 }
